# docs/frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/printer/thermal-pdf": {
      "post": {
        "operationId": "PrinterGatewayController_thermalPdf",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrintOrdersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Termal etiketka PDF (100×60mm, Gainscha GS-2408D)",
        "tags": [
          "Printer"
        ]
      }
    },
    "/printer/receipt": {
      "post": {
        "operationId": "PrinterGatewayController_receipt",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrintOrdersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A4 chek (12 ta/varaq), brauzerda avto-print",
        "tags": [
          "Printer"
        ]
      }
    },
...
  },
  "info": {
    "title": "Elchi API Gateway",
    "description": "API Gateway docs for all microservice routes",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PrintOrdersDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}